Allow to trigger autotype externally
This also forwards the necessary metadata for window tagging, so that
the popup can automatically filter the password list based on data such
as for example the window title.
In theory Arcan can also provide much more reliable data than just the
window title, for example the user could add metadata to targets to
associate some password account directly to an application.